+++ ntex-grpc/src/client/transport.rs
@@ -1,73 +1,28 @@
-use std::{cell::RefCell, convert::TryFrom, future::Future, rc::Rc, str::FromStr};
+use std::{cell::RefCell, convert::TryFrom, str::FromStr};
 
 use async_trait::async_trait;
 use ntex_bytes::{Buf, BufMut, Bytes, BytesMut};
-use ntex_connect::{Address, Connect, ConnectError, Connector as DefaultConnector};
 use ntex_h2::{self as h2, client, frame::StreamId, Stream};
 use ntex_http::{header, HeaderMap, Method, StatusCode};
-use ntex_io::{IoBoxed, OnDisconnect};
-use ntex_service::{fn_service, Service};
-use ntex_util::{channel::oneshot, future::Ready, HashMap};
+use ntex_util::{channel::oneshot, HashMap};
 
-use crate::request::{RequestContext, Response};
-use crate::service::{ClientInformation, MethodDef, Transport};
+use crate::service::MethodDef;
 use crate::{consts, utils::Data, DecodeError, GrpcStatus, Message, ServiceError};
 
-#[derive(thiserror::Error, Debug)]
-pub enum ClientError {
-    #[error("{0}")]
-    Http(#[from] ntex_h2::client::ClientError),
+use super::request::{RequestContext, Response};
+use super::{Client, Transport};
+
+pub(super) struct Inner {
+    pub(super) client: client::Client,
+    pub(super) inflight: RefCell<HashMap<StreamId, Inflight>>,
 }
 
-#[derive(Clone)]
-pub struct Client(Rc<Inner>);
-
-impl Drop for Client {
-    fn drop(&mut self) {
-        // one for current client and one for Client::start() call
-        if Rc::strong_count(&self.0) <= 2 {
-            self.0.client.close()
-        }
-    }
-}
-
-struct Inner {
-    client: client::Client,
-    inflight: RefCell<HashMap<StreamId, Inflight>>,
-}
-
-struct Inflight {
+pub(super) struct Inflight {
     _stream: Stream,
     data: Data,
     status: Option<StatusCode>,
     headers: Option<HeaderMap>,
     tx: oneshot::Sender<Result<(Option<StatusCode>, Bytes, HeaderMap, HeaderMap), ServiceError>>,
-}
-
-impl Client {
-    #[inline]
-    /// Gracefully close connection
-    pub fn close(&self) {
-        self.0.client.close()
-    }
-
-    #[inline]
-    /// Check if connection is closed
-    pub fn is_closed(&self) -> bool {
-        self.0.client.is_closed()
-    }
-
-    #[inline]
-    /// Notify when connection get closed
-    pub fn on_disconnect(&self) -> OnDisconnect {
-        self.0.client.on_disconnect()
-    }
-
-    #[inline]
-    /// Get reference to h2 client
-    pub fn get_ref(&self) -> &client::Client {
-        &self.0.client
-    }
 }
 
 #[async_trait(?Send)]
@@ -143,7 +98,7 @@ impl<T: MethodDef> Transport<T> for Client {
 }
 
 impl Inner {
-    fn handle_message(&self, mut msg: h2::Message) -> Result<(), ()> {
+    pub(super) fn handle_message(&self, mut msg: h2::Message) -> Result<(), ()> {
         let id = msg.id();
         let mut inner = self.inflight.borrow_mut();
 
@@ -229,85 +184,5 @@ impl Inner {
             }
         }
         Ok(())
-    }
-}
-
-pub struct Connector<A, T>(Rc<client::Connector<A, T>>);
-
-impl<A, T> Connector<A, T>
-where
-    A: Address,
-{
-    /// Create new grpc connector
-    pub fn new(connector: client::Connector<A, T>) -> Connector<A, T> {
-        Connector(Rc::new(connector))
-    }
-}
-
-impl<A, T> Clone for Connector<A, T>
-where
-    A: Address,
-{
-    fn clone(&self) -> Self {
-        Self(self.0.clone())
-    }
-}
-
-impl<A> Default for Connector<A, DefaultConnector<A>>
-where
-    A: Address,
-{
-    fn default() -> Self {
-        Connector::new(client::Connector::default())
-    }
-}
-
-impl<A, T> From<client::Connector<A, T>> for Connector<A, T>
-where
-    A: Address,
-    T: Service<Connect<A>, Error = ConnectError>,
-    IoBoxed: From<T::Response>,
-{
-    fn from(connector: client::Connector<A, T>) -> Self {
-        Self(Rc::new(connector))
-    }
-}
-
-impl<A, T> Connector<A, T>
-where
-    A: Address,
-    T: Service<Connect<A>, Error = ConnectError>,
-    IoBoxed: From<T::Response>,
-{
-    /// Connect and create client instance
-    pub fn create<C: ClientInformation<Client>>(
-        &self,
-        address: A,
-    ) -> impl Future<Output = Result<C, ClientError>> {
-        let fut = self.connect(address);
-
-        async move { Ok(C::create(fut.await?)) }
-    }
-
-    /// Connect to http2 server
-    pub fn connect(&self, address: A) -> impl Future<Output = Result<Client, ClientError>> {
-        let slf = self.0.clone();
-        async move {
-            let con = slf.connect(address).await?;
-            let inner = Rc::new(Inner {
-                client: con.client(),
-                inflight: RefCell::new(HashMap::default()),
-            });
-
-            let tr = inner.clone();
-            ntex_rt::spawn(async move {
-                let _ = con
-                    .start(fn_service(move |msg: h2::Message| {
-                        Ready::from(tr.handle_message(msg))
-                    }))
-                    .await;
-            });
-            Ok(Client(inner))
-        }
     }
 }