+++ ntex-grpc/src/error.rs
@@ -1,10 +1,8 @@
-use ntex_bytes::{ByteString, Bytes};
+use ntex_bytes::Bytes;
 use ntex_h2::{OperationError, StreamError};
-use ntex_http::{HeaderMap, StatusCode};
+use ntex_http::{error::Error as HttpError, HeaderMap, StatusCode};
 
-pub use ntex_http::error::Error as HttpError;
-
-pub use crate::encoding::DecodeError;
+use crate::encoding::DecodeError;
 use crate::status::GrpcStatus;
 
 #[derive(thiserror::Error, Debug)]
@@ -25,16 +23,6 @@ pub enum ServiceError {
     UnexpectedEof(Option<StatusCode>, HeaderMap),
     #[error("Grpc status {0:?}, headers: {1:?}")]
     GrpcStatus(GrpcStatus, HeaderMap),
-}
-
-#[derive(thiserror::Error, Clone, Debug)]
-pub enum ServerError {
-    #[error("{0}")]
-    Decode(#[from] DecodeError),
-    #[error("Service method is not found: {0}")]
-    NotFound(ByteString),
-    #[error("Service method is not implemented: {0}")]
-    NotImplemented(ByteString),
 }
 
 impl From<HttpError> for ServiceError {