+++ prost-build/src/code_generator.rs
@@ -203,7 +203,7 @@ impl<'a> CodeGenerator<'a> {
 
             has_fields = true;
             write.push_str(&format!(
-                "::ntex_grpc::NativeType::serialize(&self.{}, {}, ::ntex_grpc::DefaultValue::Default, dst);",
+                "::ntex_grpc::NativeType::serialize(&self.{}, {}, ::ntex_grpc::types::DefaultValue::Default, dst);",
                 to_snake(field.name()),
                 field.number()
             ));
@@ -215,7 +215,7 @@ impl<'a> CodeGenerator<'a> {
                 to_snake(field.name()),
             ));
             encoded_len.push_str(&format!(
-                " + ::ntex_grpc::NativeType::serialized_len(&self.{}, {}, ::ntex_grpc::DefaultValue::Default)",
+                " + ::ntex_grpc::NativeType::serialized_len(&self.{}, {}, ::ntex_grpc::types::DefaultValue::Default)",
                 to_snake(field.name()),
                 field.number()
             ));
@@ -250,7 +250,7 @@ impl<'a> CodeGenerator<'a> {
 
             has_fields = true;
             write.push_str(&format!(
-                "::ntex_grpc::NativeType::serialize(&self.{}, 0, ::ntex_grpc::DefaultValue::Default, dst);",
+                "::ntex_grpc::NativeType::serialize(&self.{}, 0, ::ntex_grpc::types::DefaultValue::Default, dst);",
                 to_snake(oneof.name()),
             ));
             read.push_str(&format!(
@@ -265,7 +265,7 @@ impl<'a> CodeGenerator<'a> {
                 to_snake(oneof.name()),
             ));
             encoded_len.push_str(&format!(
-                " + ::ntex_grpc::NativeType::serialized_len(&self.{}, 0, ::ntex_grpc::DefaultValue::Default)",
+                " + ::ntex_grpc::NativeType::serialized_len(&self.{}, 0, ::ntex_grpc::types::DefaultValue::Default)",
                 to_snake(oneof.name()),
             ));
             default.push_str(&format!(
@@ -513,7 +513,7 @@ impl<'a> CodeGenerator<'a> {
         self.depth += 1;
         for (field, idx) in &fields {
             write.push_str(&format!(
-                "{}::{}(ref value) => ::ntex_grpc::NativeType::serialize(value, {}, ::ntex_grpc::DefaultValue::Default, dst),",
+                "{}::{}(ref value) => ::ntex_grpc::NativeType::serialize(value, {}, ::ntex_grpc::types::DefaultValue::Default, dst),",
                 to_upper_camel(oneof.name()),
                 to_upper_camel(field.name()),
                 field.number()
@@ -526,7 +526,7 @@ impl<'a> CodeGenerator<'a> {
                 field.number(),
             ));
             encoded_len.push_str(&format!(
-                "{}::{}(ref value) => ::ntex_grpc::NativeType::serialized_len(value, {}, ::ntex_grpc::DefaultValue::Default),",
+                "{}::{}(ref value) => ::ntex_grpc::NativeType::serialized_len(value, {}, ::ntex_grpc::types::DefaultValue::Default),",
                 to_upper_camel(oneof.name()),
                 to_upper_camel(field.name()),
                 field.number()
@@ -567,7 +567,7 @@ impl<'a> CodeGenerator<'a> {
         self.buf.push_str(&format!(
             "
             /// Encodes the message to a buffer.
-            fn serialize(&self, _: u32, _: ::ntex_grpc::DefaultValue<&Self>, dst: &mut ::ntex_grpc::BytesMut) {{
+            fn serialize(&self, _: u32, _: ::ntex_grpc::types::DefaultValue<&Self>, dst: &mut ::ntex_grpc::BytesMut) {{
                 match *self {{ {} }}
             }}",
             write
@@ -584,7 +584,7 @@ impl<'a> CodeGenerator<'a> {
         self.buf.push_str(&format!(
             "
             /// Returns the encoded length of the message without a length delimiter.
-            fn serialized_len(&self, _: u32, _: ::ntex_grpc::DefaultValue<&Self>) -> usize {{
+            fn serialized_len(&self, _: u32, _: ::ntex_grpc::types::DefaultValue<&Self>) -> usize {{
                 match *self {{
                     {}
                 }}